Improved rendering of external content in news items
News content was rendered in a WebView from the HTML present
in a JSON response.This caused issues with images and videos which
were served over HTTPS, as the content was not directly loaded.

Now, the WebView calls a new endpoint (already deployed in backend)
which serves the news contents as pure HTML.

This fixes all issues with external content.