Fixed download bar not dissappearing when finished in background
If you downloaded an update and turned the screen off / exited the
app, the app would still show that it's downloading even though it
was already done. This was because the downloadListener disconnects
and does not query its initial status when reconnecting.

The method to obtain initial status has ben extended to also support
getting completion / verification failed status.