Users can now contribute by submitting update file names
All app users can opt-in to submit update file names found in the .Ota
folder to our admins, so we can import new updates faster and more
easily.

The opt-in is implemented as following:
- Users updating from version 2.6.3 or earlier see a prompt to
contribute when they open the new version of the app for the 1st time
- Users installing the app for the first time can opt-in on the last
page of the setup wizard
- All users can always opt-in or opt-out from the Settings screen

The file check runs every 15 minutes (platform minimum interval) and
only scans if there is a network connection. If the .Ota folder does
not exist nothing is scanned. If a file is found, we check if it has
already been submitted by the user before, otherwise we submit it and
show a "thank you" notification.