Hacked-in support for the OnePlus 7 Pro Series
The OnePlus 7 Pro Series uses new build properties for the
human-readable operating system and the region-specific device model.
But these are conflicting with older devices. So i made the ugly choice
to add an exception for this device only, hoping that OnePlus reverts
this decision in their next device model.