Release v4.0.0 - If the user has selected a different device or update method, the "changelog" dialog will now show a small notice informing them about this - If a download fails, the app now shows a download link, so that the user can manually download it via the browser - The app now shows the MD5 hash of the update ZIP, below the file name - Greatly improved UI/UX for server banners - they're now behind a BottomSheet, that's shown after the user taps on it. Thus, it frees up space on the main update page. - WebViews within the app now have improved error handling, and do not show a blank white screen on first load anymore - Added a loading shimmer animation in many views - Switched to Google Sans for several title/header TextViews - Several TextViews are now selectable - Added a statistical banner in NewsFragment, that shows the number of unread articles. This is coupled with a badge that's displayed on the "News" tab. Tapping on this banner toggles between showing only unread articles, and all articles. - Added error/empty states to convey useful information to the user - Moved to Retrofit for network requests, and Glide for image requests
This version (and higher) support Google Play's in-app updates feature - meaning the user won't have to open Play Store to update the app, they can do so from within the app itself.
Currently, the app checks for an update once every 2 days, and the user can choose to accept or reject an update via the dialog shown by Play Services. In case an update is rejected 7 times, or if the update is 14 days old, the app shows a fullscreen UI (again, served by Play Services) to inform the user that this is an "important" update, and should be installed.

Additionally, major redesigns were done:
- News article page
- Install guide page (for both rooted and non-rooted devices)
- Setup/onboarding page
- About this app page
- Most dialogs were switched to BottomSheets instead, as it offers better UX

The main goal of the redesign was to refresh everything upto modern standards, and conform to our "brand identity". The previous UI/UX of the app looked unpolished, unprofessional, etc.
As part of this redesign, we epdated app logo and other icons to their "rounded" variants. Rounder icons convey a sense of friendliness, reliability, safety, etc. Since this app is made by the community, for the community, it should reflect the same personality/character traits of that community - friendly, supportive, dependable.
If you're a UX designer, you'd probably understand this change right away - if not, check out this article: https://uxplanet.org/rounded-or-sharp-corner-buttons-def3977ed7c4.