Fixed incorrect formatting of file sizes We were using android.text.format.Formatter.formatFileSize() for formatting bytes to a more human-readable string. That function formats using SI units on Oreo and above, but IEC units pre-Oreo. Since admin portal uses IEC units (MiB instead of MB), we needed to add code to normalize behaviour across API levels.
Also affected in this commit is the "Amount of RAM" row in the "Device" tab. It used to display RAM in IEC units, and only in MB. This has now been changed to reflect RAM in SI units, which is what the average user expects.

Note:, the formatter automatically chooses between MB/GB/etc, which is a slight improvement over the previous implementation: sizes were only shown in MB earlier.