Got rid of ContributorActivity: converted to bottom sheet Some dialogs and snackbars are now initialized lazily to improve performance and memory usage
Also fixed a bug in `BottomSheetDialog` layouts where the text inside `NestedScrollView` overlapped with other views above and below it. This happened only if the available screen space wasn't enough to contain all views (which is why the text was wrapped in a `NestedScrollView`). However due to how `ConstraintLayout` works, the scroll view wasn't filling remaining space properly.