Migrated from deprecated AdMob interstitial APIs in preparation of v20 https://developers.google.com/admob/android/migration
This commit also includes an interstitial advertising unit ID change, because from v5.2.0 onwards (of the app), we'll rely on AdMob's built-in frequency capping instead of a custom SharedPreferences-based solution (save last shown timestamp and limit showing based on that) that probably was buggy (some users reported issues, though they never provided screen records to confirm the issue).

Frequency will remain the same: limit interstitials to be shown only once in a 5-minute window. Because this AdMob-controlled capping will conflict with the older SharedPreferences-based capping, we changed the advertising unit ID to avoid conflicts.