fix(news-item): net::ERR_CONNECTION_REFUSED even if article could load
Culprit: WebView successfully loaded the article, but also reported a "net::ERR_CONNECTION_REFUSED" error for scripts in the main URL's page (e.g. Google AdSense and Cloudflare Web Analytics). This happened only for a few users, and it's not clear why it was not experienced by everyone.
Since the contents of the article itself aren't affected by these third-party scripts, the fix is to ignore errors that aren't meant for our own API URLs.

Also improved UX for these errors: it's a dismissable dialog now, instead of a full-layout UI.
This keeps WebView visible, so that if the app has a similar bug in the future, users can simply dismiss the dialog and continue reading the article.