feat: device images are now loaded from the server
This is a stop-gap feature that constructs an image URL based on device name. There's a low probability it won't work for all devices, since the server may not have the image with the exact name as constructed.

It'll be improved in a future API version (v2.7), which will return image URLs as part of the network request to retrieve the device list.

As part of this change, all bundled images except oneplus7pro.webp (it's used as a fallback) in drawable-nodpi have been deleted. This would reduce APK size by a few hundred KBs.