feat(in-app-update): remove "at most once every 2 days" restriction
Our priority is to have users be on the latest app version, as it always comes with new features and bug fixes.

We've always checked for app updates every time the user opens the app, but earlier we requested the user to install it only if at least 2 days had passed since the last time we knew there was an app update available.

This convoluted logic mostly just resulted in an unnecessary 2-day "silence" period.