feat(fcm): introduce device-only topic in addition to device+method
We often need to send notifications to a large group of devices, regardless of the selected update method.
Given how many device/method combinations there are, it doesn't make sense to send notifications to num(devices) * num(methods) topics in total.

With this commit we're cutting it by a factor of ~10, though it won't be used until much later when most people have updated to a version of the app with this feature.