perf(NewsList/scroll): avoid using `layoutInfo` in composables
`layoutInfo` is marked as a `@FrequentlyChangingValue`, and is updated
after every scroll or remeasure. Using it in a `@Composable` function
directly could cause recompositions on every change, thus costing a lot
in terms of performance (incl. a possible infinite recomposition loop).

It's important to note that the original source did not use `layoutInfo`
anywhere. The change must've been for convenience; however, lint did not
warn of `layoutInfo` usage at the time. Perhaps it's a new check.