fix(main): Play in-app update snackbar no longer immediately dismisses
This snackbar shouldn't have been self-dismissable in the first place,
but the bug was mainly because of duplicating the dismissal logic in
both `AppUpdateInfo` and `FlexibleAppUpdateProgress`. When an in-app
update was in progress, the latter composable updated `snackbarText` to
`AppUpdateDownloadedSnackbarData`, but because that recomposes, control
was gained by `AppUpdateInfo`, which immediately dismissed that same
snackbar. Only an app restart showed the snackbar again, via
`AppUpdateInfo`, which effectively meant that the in-app update would've
been installed only after 2 app restarts (if user tapped on the 'Reload'
action the 2nd time).

Added comments to remind us of duplicity in the future, should we ever
change this snackbar to be self-dismissable again. Ideally, only
`FlexibleAppUpdateProgress` should have dismissal code.