feat(settings): add convenience button to clear API & image caches
Caches are useful because they allow skipping re-downloading unchanged
resources, thus also saving the user's network bandwidth (though we only
have minimal network response sizes).

However, sometimes for debugging or troubleshooting purposes, clearing
cache often solves the issue of stale data being served.

While the user could always clear local cache via the "App info" section
in Android, they can only do so after several clicks, including a long
press. So, this makes it much easier because they don't have to leave
the app to clear cache now.

The intention is that this will only be used when directed by us, or if
the user is technically proficient enough to know that a cache clear
might be helpful. Unfortunately, if you give people a button, they're
gonna click it, so hopefully this will not be abused & clicked regularly,
because that would just defeat the purpose of the cache. Still, more
power to the user.